third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-parse-event-stream-expected.txt

[Worker] Test EventSource text/event-stream parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-parse-event-stream.js
PASS [Worker] es = new EventSource("resources/event-stream.php"); did not throw exception.
FAIL [Worker] count should be 7. Was -1.
[Worker] DONE
PASS successfullyParsed is true

TEST COMPLETE
